backend/database.py
import os
import mysql.connector
from mysql.connector import Error

def get_db_connection():
    try:
        conn = mysql.connector.connect(
            host=os.getenv("DB_HOST", "mysql"),
            user=os.getenv("DB_USER", "user"),
            password=os.getenv("DB_PASSWORD", "password"),
            database=os.getenv("DB_NAME", "minemind"),
            port=3306
        )
        return conn
    except Error as e:
        print(f"Error connecting to MySQL: {e}")
        return None
