backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pandas==2.1.4
numpy==1.26.2
requests==2.31.0
mysql-connector-python==8.1.0
chromadb==0.4.6
python-dotenv==1.0.0
flask-bcrypt==1.0.1
PyJWT==2.8.0
